Декоратор delete_previous_message перенесён в src/bot/utils.py, к нему добавлен докстринг.
